requirements.txt
requests==2.13.0
python-decouple==3.0
